fetch data in applicationDidBecomeActive, not on launch
I want to check if trying to fetch after a relaunch in the background (during the night or after a crash), when the app will almost certainly be suspended a second later and won’t finish the request, isn’t actually doing more harm than good. So now we’ll be fetching only when the app is launched into the foreground, not in the background.